requirements.txt
streamlit
pandas
yfinance
plotly
gspread
google-auth
matplotlib